Rename component to ItemIdSearchTextarea
